requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
appdirs==1.4.4
    # via -r requirements.in
biopython==1.81
    # via -r requirements.in
certifi==2023.5.7
    # via requests
charset-normalizer==3.3.2
    # via requests
idna==3.4
    # via requests
numpy==1.24.3
    # via
    #   biopython
    #   pandas
pandas==2.1.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
pytz==2023.3.post1
    # via pandas
requests==2.29.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
tzdata==2023.3
    # via pandas
urllib3==1.26.18
    # via requests
